validator-plugin-cnig - DatabaseSUP - move legacy DatabaseSUPFactory to DatabaseTestFactory in tests (refs #176)
