+++ validator-plugin-cnig/src/main/java/fr/ign/validator/cnig/CnigPlugin.java
@@ -4,13 +4,13 @@ import fr.ign.validator.Context;
 import fr.ign.validator.cnig.process.CreateShapefilesPostProcess;
 import fr.ign.validator.cnig.process.CustomizeIdurbaPreProcess;
 import fr.ign.validator.cnig.process.DocUrbaComPostProcess;
-import fr.ign.validator.cnig.process.SupRelationsPostProcess;
 import fr.ign.validator.cnig.process.DocUrbaPostProcess;
 import fr.ign.validator.cnig.process.PerimetreScotPostProcess;
+import fr.ign.validator.cnig.process.SupRelationsPostProcess;
 import fr.ign.validator.cnig.validation.attribute.InseeValidator;
 import fr.ign.validator.cnig.validation.document.AtLeastOneWritingMaterialValidator;
-import fr.ign.validator.cnig.validation.metadata.CnigMetadataKeywordsValidator;
 import fr.ign.validator.cnig.validation.metadata.CnigMetadataDateOfLastRevisionValidator;
+import fr.ign.validator.cnig.validation.metadata.CnigMetadataKeywordsValidator;
 import fr.ign.validator.cnig.validation.metadata.CnigMetadataReferenceSystemIdentifierValidator;
 import fr.ign.validator.cnig.validation.metadata.CnigSpecificationsValidator;
 import fr.ign.validator.cnig.validation.metadata.CnigTypeValidator;
