+++ validator-plugin-cnig/src/test/java/fr/ign/validator/cnig/sup/DatabaseSUPDuplicatedTest.java
@@ -38,7 +38,7 @@ public class DatabaseSUPDuplicatedTest {
 
     @Before
     public void setUp() throws Exception {
-        DatabaseSUPFactory factory = new DatabaseSUPFactory(folder.getRoot());
+        TestDatabaseFactory factory = new TestDatabaseFactory(folder.getRoot());
         File dataDirectory = ResourceHelper.getResourceFile(getClass(), "/jointure_sup/duplicated/DATA");
         db = factory.createFromDataDirectory(dataDirectory);
     }
